:recycle: PointService 인터페이스로 수정, 기존 PointService PointServiceImpl 로 수정
- 추후 확정성을 위해 기존 클래스 였던 PointService 를 인터페이스로 변환하고 PointServiceImpl 구현체를 DI 하도록 변경
