+++ src/main/java/io/hhplus/tdd/point/PointServiceImpl.java
@@ -1,0 +1,19 @@
+package io.hhplus.tdd.point;
+
+import io.hhplus.tdd.database.PointHistoryTable;
+import io.hhplus.tdd.database.UserPointTable;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class PointServiceImpl implements PointService {
+
+    private final UserPointTable userPointTable;
+    private final PointHistoryTable pointHistoryTable;
+
+    @Override
+    public UserPoint searchUserPoint(long userId) {
+        return userPointTable.selectById(userId);
+    }
+}
